+++ after/ftplugin/lua.lua
@@ -1,0 +1,17 @@
+-- Vim filetype plugin file
+--
+-- It is in Lua because Lua!
+--
+-- Language:	Lua
+-- Maintainer:	Theo P.
+-- Last Change:	2023-10-10
+
+local indent = 2
+
+vim.opt_local.expandtab = true
+vim.opt_local.tabstop = indent
+vim.opt_local.softtabstop = indent
+vim.opt_local.shiftwidth = indent
+vim.opt_local.colorcolumn = "120"
+vim.opt_local.textwidth = 119
+