+++ init.lua
@@ -1,25 +1,29 @@
---[[ init.lua
---
---      \/       \/
---      /\_______/\
---     /   o   o   \
---    (  ==  ^  ==  )
---     )           (
---    (             )
---    ( (  )   (  ) )
---   (__(__)___(__)__)
---  ___
---   | |_  _  _     o __
---   | | |(/_(_)\_/ | |||
---
--- Cat : https://www.asciiart.eu/animals/cats
---      My friend added a few layers of belly so it looks more like chunky my cat Oliver
--- Logo: $ figlet -f mini Theovim
---
--- Initialize all configuration files
---]]
+--- init.lua
+---
+---      \/       \/
+---      /\_______/\
+---     /   o   o   \
+---    (  ==  ^  ==  )
+---     )           (
+---    (             )
+---    ( (  )   (  ) )
+---   (__(__)___(__)__)
+---  ___
+---   | |_  _  _     o __
+---   | | |(/_(_)\_/ | |||
+---
+--- Cat : https://www.asciiart.eu/animals/cats
+---      My friend added a few layers of belly so it looks more like chunky my cat Oliver
+--- Logo: $ figlet -f mini Theovim
+---
+--- Initialize all configuration files
 
--- Try catch for modules
+--- safe_require()
+--- Try calling `require` for the given module.
+--- If unsuccessful, print the error message using `vim.notify()`
+---
+---@param module string a name of the module to `require`
+---@return unknown module from `pcall` if the call was successful, otherwise nil
 local function safe_require(module)
   local status, loaded_module = pcall(require, module)
   if status then
@@ -29,36 +33,32 @@ local function safe_require(module)
   return nil
 end
 
--- User configuration
-safe_require("config")
-
 -- Core config modules
 safe_require("core")
 safe_require("plugins")
 
+-- Telescope and Treesitter
+safe_require("tele")
+safe_require("ts")
+
+-- LSP
+safe_require("lsp")
+
 -- Theovim built-in UI elements
-local highlights = safe_require("ui.highlights")
-if highlights then highlights.setup() end
+safe_require("ui")
 
-local statusline = safe_require("ui.statusline")
-if statusline then statusline.setup() end
-local tabline = safe_require("ui.tabline")
-if tabline then tabline.setup() end
-local winbar = safe_require("ui.winbar")
-if winbar then winbar.setup() end
-local dashboard = safe_require("ui.dashboard")
-if dashboard then dashboard.setup() end
+-- Theo's Neovide settings
+if vim.g.neovide then
+  local padding = 10
+  vim.g.neovide_padding_top = padding
+  vim.g.neovide_padding_bottom = padding
+  vim.g.neovide_padding_right = padding
+  vim.g.neovide_padding_left = padding
 
-local notepad = safe_require("ui.notepad")
-if notepad then notepad.setup() end
+  vim.g.neovide_hide_mouse_when_typing = true
+  vim.g.neovide_cursor_vfx_mode = "railgun"
 
--- LSP configurations
-safe_require("lsp.lsp")
-safe_require("lsp.completion")
-
--- Plugin configurations
-safe_require("config.fuzzy")
-safe_require("config.treesitter")
-
--- Other Theovim features
-safe_require("misc")
+  vim.g.neovide_transparency = 0.69
+  vim.g.neovide_floating_blur_amount_x = 2.0
+  vim.g.neovide_floating_blur_amount_y = 2.0
+end